requirements.txt
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.28.1
pygame>=2.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
pytest>=7.3.1 